pyproject.toml
[tool.poetry]
name = "tplink"
version = "0.1.0"
description = ""
authors = ["Vitor Gabriel <edvitor13@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-kasa = "^0.6.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
